EBGURG-471 resolve review comments

+++ vision/agv_speed_calculator/src/agv_speed_calculator/AgvConstants.hpp
@@ -21,7 +21,7 @@ const uint8_t M2 = 6;
 // The speed for the servo's
 const uint8_t gServoSpeed_pwm = 100; // Was 63
 
-const uint32_t gMinAmountMeasurements = 2;
+const uint16_t gMinAmountMeasurements = 2;
 // The length beween the lines to be detected
 const double gMeasurementLength_m = 0.5;
 // Amount of ms the tracker sensor needs to return true for the signal to be
